Extract shared ChildProcessSpawner tests
