+++ packages/effect/test/unstable/process/fixtures/bash/no-permissions.sh
