+++ packages/effect/test/unstable/process/fixtures/bash/parent-exits-early.sh
