+++ packages/effect/test/unstable/process/fixtures/bash/spawn-children.sh
