+++ packages/effect/test/unstable/process/fixtures/config/SHOUTING
