+++ packages/effect/test/unstable/process/fixtures/config/nested/config
